Refactor StringOrData type
Before the `StringOrData` type alway represented a string and automatically converted any data (byte slice) to a string using utf8 encoding, which lead to potentially lost data.
Now the `StringOrData` type is an enum, which holds a reference to the original value and can be converted to either a byte slice, also referencing to the original value, or
a Cow String, which either holds a reference to the original value or creates a new owned one. The reason for this is, that in the conversion from an data type to a string,
utf8 encoding is used. There might be some invalid utf8 data in the type, which are lost in the conversion, therefore we have to create a new owned string from that.